Add new UI

+++ palmatum/internal/core/sites.go
@@ -1,6 +1,7 @@
 package core
 
 import (
+	"database/sql"
 	"errors"
 	"fmt"
 	"github.com/codemicro/palmatum/palmatum/internal/database"
@@ -78,8 +79,10 @@ func (c *Core) DeleteSite(siteSlug string) error {
 		return fmt.Errorf("commit transaction: %w", err)
 	}
 
-	if err := os.Remove(c.getPathOnDisk(contentPath)); err != nil {
-		return fmt.Errorf("remove path: %w", err)
+	if contentPath != "" {
+		if err := os.Remove(c.getPathOnDisk(contentPath)); err != nil {
+			return fmt.Errorf("remove path: %w", err)
+		}
 	}
 
 	if err := c.BuildKnownRoutes(); err != nil {
@@ -99,11 +102,24 @@ func (c *Core) UpdateSite(s *database.SiteModel) error {
 
 var (
 	ErrInvalidDomain  = newError("invalid domain")
-	ErrInvalidPath    = newError("invalid path")
+	ErrInvalidPath    = newError("invalid path (must start with /)")
 	ErrRouteNotUnique = newError("route is not unique")
 )
 
 func (c *Core) CreateRoute(siteSlug, domain, path string) (*database.RouteModel, error) {
+	tx, err := c.Database.Beginx()
+	if err != nil {
+		return nil, fmt.Errorf("begin database transaction: %w", err)
+	}
+	defer tx.Rollback()
+
+	if _, err := database.GetSite(tx, siteSlug); err != nil {
+		if errors.Is(err, sql.ErrNoRows) {
+			return nil, ErrInvalidSlug
+		}
+		return nil, fmt.Errorf("get site from database: %w", err)
+	}
+
 	domain = strings.TrimSpace(domain)
 
 	if domain == "" {
@@ -118,7 +134,7 @@ func (c *Core) CreateRoute(siteSlug, domain, path string) (*database.RouteModel,
 
 	var id int
 
-	if err := c.Database.QueryRowx("INSERT INTO routes(site, domain, path) VALUES (?, ?, ?) RETURNING id", siteSlug, domain, path).Scan(&id); err != nil {
+	if err := tx.QueryRowx("INSERT INTO routes(site, domain, path) VALUES (?, ?, ?) RETURNING id", siteSlug, domain, path).Scan(&id); err != nil {
 		var e sqlite3.Error
 		if errors.As(err, &e) {
 			if e.ExtendedCode == sqlite3.ErrConstraintForeignKey {
@@ -129,6 +145,10 @@ func (c *Core) CreateRoute(siteSlug, domain, path string) (*database.RouteModel,
 			}
 		}
 		return nil, fmt.Errorf("call database: %w", err)
+	}
+
+	if err := tx.Commit(); err != nil {
+		return nil, fmt.Errorf("commit databse transaction: %w", err)
 	}
 
 	if err := c.BuildKnownRoutes(); err != nil {
